Commit adding new and modern project
- Actually divided the code between the folder so it looks more organized
- Runs perfectly (or just like last time)
